Add workaround for displaying ellipsis on windows
On windows, ellipses are usually outside the chars which can be displayed in our
way to display plain text /with  `capture.output(print(...))` in the current
locale. The leads to strangely escaped chars and we would have such a escaped
chars each time we display a dataframe.

This problem is probably not going away soon:
https://github.com/IRkernel/repr/issues/28#issuecomment-208574856

Work around this by simple using three dots.