Suppress stacktraces in repr_text
Stacktraces in repr_text look bad when packages like ggplot uses the included
print to build a plot but fail due to some situation signaled by stop(). The old
result was that a user error resulted in a stacktrace deep in our repr stack. The
new result is that only the error message is shown.